refactor: split commands/bug.rs into per-action submodules
Promotes commands/bug.rs (1297 lines) to commands/bug/ with one
handler per file (list, view, history, search, my, create, clone,
update, shared). Drops file-level cognitive complexity from 54 and
splits previously-shared test fixtures alongside the code they cover.

shared.rs is currently a placeholder: the Sonar-flagged duplications
were almost entirely test fixtures (mock JSON payloads, BugAction
constructors), which now live in each handler's own mod tests. No
cross-handler production logic was mechanically extractable per the
spec's discipline rule.

Public path crate::commands::bug::execute unchanged.
